// src/api/models/createProfileDto.ts
/**
 * Generated by orval v6.10.2 🍺
 * Do not edit manually.
 * API
 * API DOCS
 * OpenAPI spec version: 1.0
 */
import type { CreateProfileDtoUserType } from './createProfileDtoUserType';
import type { BusinessDetails } from './businessDetails';
import type { PrismaConnect } from './prismaConnect';

export interface CreateProfileDto {
  id?: string;
  createdAt?: string;
  firebaseUID?: string;
  email?: string;
  phone?: string;
  fullName?: string;
  userType?: CreateProfileDtoUserType;
  birthday?: string;
  businessDetails?: BusinessDetails;
  bio?: string;
  address?: string;
  createdById?: string;
  createdByCollection?: string;
  password: string;
  products: PrismaConnect;
}